+++ RedPandaIDE/widgets/functiontipwidget.cpp
@@ -1,0 +1,13 @@
+#include "functiontipwidget.h"
+
+#include <QHBoxLayout>
+
+FunctionTipWidget::FunctionTipWidget(QWidget *parent) : QWidget(parent)
+{
+    setWindowFlags(Qt::ToolTip);
+
+    mLabel = new QLabel(this);
+    mLabel->setText("Test");
+    this->setLayout(new QHBoxLayout());
+    this->layout()->addWidget(mLabel);
+}